Refactor the llama example to make it more in sync with the other ones.
